+++ buildSrc/src/main/kotlin/ktx/Versions.kt
@@ -1,6 +1,6 @@
 package ktx
 
-const val gdxVersion = "1.13.1"
+const val gdxVersion = "1.13.5"
 const val kotlinCoroutinesVersion = "1.10.1"
 
 const val artemisOdbVersion = "2.3.0"