+++ tiled/src/test/kotlin/ktx/tiled/MapLayerTest.kt
@@ -2,6 +2,8 @@ package ktx.tiled
 
 import com.badlogic.gdx.maps.MapLayer
 import com.badlogic.gdx.maps.MapLayers
+import com.badlogic.gdx.maps.tiled.TiledMapTileLayer
+import com.badlogic.gdx.maps.tiled.TiledMapTileLayer.Cell
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertFalse
 import org.junit.Assert.assertNull
@@ -84,5 +86,25 @@ class MapLayerTest {
     val actual = MapLayers()
 
     assertFalse(actual.isNotEmpty())
+  }
+
+  @Test
+  fun `should iterate over each cell in the TiledMapTileLayer`() {
+    val layer = TiledMapTileLayer(2, 2, 32, 32).apply {
+      setCell(0, 0, Cell())
+      setCell(1, 0, null) // this cell is skipped during iteration
+      setCell(0, 1, Cell())
+      setCell(1, 1, Cell())
+    }
+
+    val collectedCells = mutableListOf<Triple<Int, Int, Cell>>()
+    layer.forEachCell { cell, cellX, cellY ->
+      collectedCells.add(Triple(cellX, cellY, cell))
+    }
+
+    assertEquals(3, collectedCells.size)
+    assertEquals(Triple(0, 0, layer.getCell(0, 0)), collectedCells[0])
+    assertEquals(Triple(0, 1, layer.getCell(0, 1)), collectedCells[1])
+    assertEquals(Triple(1, 1, layer.getCell(1, 1)), collectedCells[2])
   }
 }